adding timetable

+++ backend/src/main/java/com/netcracker/edu/backend/entity/Subjects.java
@@ -15,8 +15,10 @@ import java.util.Set;
 public class Subjects {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
+    
+    private int id;
     private String subject;
+    
     private int teacherId;
     /*@JsonIgnore
     private Set<Timetable> timetablesById;*/
@@ -32,11 +34,11 @@ public class Subjects {
     
     @Id
     @Column(name = "id", nullable = false)
-    public long getId() {
+    public int getId() {
         return id;
     }
     
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
